requirements.txt
arabic-reshaper==3.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
deep-translator==1.11.4
evdev==1.9.2
gTTS==2.5.4
idna==3.11
playsound==1.3.0
pygame==2.6.1
pynput==1.8.1
pyperclip==1.11.0
python-bidi==0.6.7
python-xlib==0.33
requests==2.32.5
six==1.17.0
soupsieve==2.8
termcolor==3.1.0
typing_extensions==4.15.0
urllib3==2.5.0
